requirements.txt
PyMySQL==0.10.1
psycopg2-binary
pathlib
PyYAML==5.4.1
datetime
mysql-replication